services/video-worker/requirements.txt
fastapi==0.115.7
uvicorn==0.34.0
httpx==0.28.1
pydantic==2.10.6
Pillow==11.1.0
playwright==1.50.0
python-dotenv==1.0.1
gtts==2.5.3
aiofiles==24.1.0
edge-tts==7.0.2

# RunPod Serverless SDK
runpod>=1.6.0

# Cloudflare R2 upload (S3-compatible)
boto3>=1.34.0

# SadTalker / Stable Diffusion deps
torch>=2.1.0
torchvision>=0.16.0
diffusers>=0.27.0
transformers>=4.38.0
accelerate>=0.27.0
huggingface_hub>=0.21.0
